refactor(World, RacingManager, RacingManagerException): 컴트롤러 내부의 로직을 racingManager로 옮김
